python/requirements.txt
matplotlib==3.4.2
numpy==1.20.3
omegaconf==2.1.1
opencv_python==4.5.5.64
Pillow==9.0.1
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.61.0
opencv-python~=4.5.5.64
scipy~=1.6.3
fvcore~=0.1.5
detectron2~=0.6